feat: move logic to store, refactor components
