+++ package.json
@@ -15,7 +15,8 @@
     "emotion": "10.0.27",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-scripts": "5.0.1"
+    "react-scripts": "5.0.1",
+    "zustand": "^4.4.7"
   },
   "scripts": {
     "start": "cross-env BROWSER=none react-scripts start",
